feat: Move AudioRepository to new folder structure
