+++ src/script/notification/NotificationRepository.test.ts
@@ -28,6 +28,7 @@ import {Runtime} from '@wireapp/commons';
 import {Availability} from '@wireapp/protocol-messaging';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
+import {AudioRepository} from 'Repositories/audio/AudioRepository';
 import {CallingRepository} from 'src/script/calling/CallingRepository';
 import {TERMINATION_REASON} from 'src/script/calling/enum/TerminationReason';
 import {ConnectionMapper} from 'src/script/connection/ConnectionMapper';
@@ -60,7 +61,6 @@ import {createUuid} from 'Util/uuid';
 
 import {NotificationRepository} from './NotificationRepository';
 
-import {AudioRepository} from '../audio/AudioRepository';
 import {CallingViewMode, CallState} from '../calling/CallState';
 import {ConnectionEntity} from '../connection/ConnectionEntity';
 import {ConversationState} from '../conversation/ConversationState';
