+++ src/script/notification/NotificationRepository.ts
@@ -28,6 +28,8 @@ import {Availability} from '@wireapp/protocol-messaging';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {AssetRepository} from 'Repositories/assets/AssetRepository';
+import {AudioRepository} from 'Repositories/audio/AudioRepository';
+import {AudioType} from 'Repositories/audio/AudioType';
 import {CallingRepository} from 'src/script/calling/CallingRepository';
 import {Declension, t, getUserName} from 'Util/LocalizerUtil';
 import {getLogger, Logger} from 'Util/Logger';
@@ -38,8 +40,6 @@ import {ValidationUtilError} from 'Util/ValidationUtil';
 
 import {PermissionState} from './PermissionState';
 
-import {AudioRepository} from '../audio/AudioRepository';
-import {AudioType} from '../audio/AudioType';
 import {CallingViewMode, CallState} from '../calling/CallState';
 import {TERMINATION_REASON} from '../calling/enum/TerminationReason';
 import type {ConnectionEntity} from '../connection/ConnectionEntity';
