+++ src/script/view_model/MainViewModel.ts
@@ -20,13 +20,13 @@
 import {container} from 'tsyringe';
 
 import type {AssetRepository} from 'Repositories/assets/AssetRepository';
+import type {AudioRepository} from 'Repositories/audio/AudioRepository';
 
 import {ActionsViewModel} from './ActionsViewModel';
 import {CallingViewModel} from './CallingViewModel';
 import {ContentViewModel} from './ContentViewModel';
 import {ListViewModel} from './ListViewModel';
 
-import type {AudioRepository} from '../audio/AudioRepository';
 import type {BackupRepository} from '../backup/BackupRepository';
 import type {CallingRepository} from '../calling/CallingRepository';
 import type {ClientRepository} from '../client';
